mysite/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
python-memcached==1.59
pytz==2018.4
six==1.11.0
